removed certUtil.exe dependency + formatting
Both Get-FileHash and certUtil.exe can be used to compute the hash value of a file. However, Get-FileHash is a built-in PowerShell command that can be used to compute the hash value of a file without relying on the external tools certUtil.exe.

Added comments and modified formatting and included license at the top.